fix(platform): make log_redact V8-compatible (drop node:crypto)
The Convex bundler rejected log_redact.ts because importing node:crypto
from a non-action module forces the whole import graph into the Node
runtime. Switched to a pure-JS FNV-1a hash so the helper runs in V8
without a 'use node' directive.

- Extract fnv1aHash from response_cache/cache_key.ts to a shared
  lib/fnv1a.ts module so both call sites consume the same impl.
- Replace SHA-256 in summarizeForLog with fnv1aHash; rename the
  returned field sha256 -> digest to stop misadvertising the algo.
- Update tests to expect the 16-char FNV digest. Plaintext-leak
  guarantee unchanged (hash is one-way; digest does not echo input).

bun run check passes (2719 tests, lint, typecheck). FNV is not
cryptographic — fine here, the only property we need is non-reversible
correlation for log diffing.
