+++ services/platform/convex/lib/__tests__/log_redact.test.ts
@@ -3,9 +3,9 @@ import { describe, it, expect } from 'vitest';
 import { summarizeForLog } from '../log_redact';
 
 describe('summarizeForLog', () => {
-  it('returns a 12-char hex sha256 prefix and the byte length', () => {
+  it('returns a 16-char hex digest and the byte length', () => {
     const result = summarizeForLog('hello world');
-    expect(result.sha256).toMatch(/^[0-9a-f]{12}$/);
+    expect(result.digest).toMatch(/^[0-9a-f]{16}$/);
     expect(result.len).toBe('hello world'.length);
   });
 
@@ -21,14 +21,14 @@ describe('summarizeForLog', () => {
   it('is deterministic for identical inputs', () => {
     const a = summarizeForLog('the quick brown fox');
     const b = summarizeForLog('the quick brown fox');
-    expect(a.sha256).toBe(b.sha256);
+    expect(a.digest).toBe(b.digest);
     expect(a.len).toBe(b.len);
   });
 
   it('produces different digests for different inputs', () => {
     const a = summarizeForLog('alpha');
     const b = summarizeForLog('beta');
-    expect(a.sha256).not.toBe(b.sha256);
+    expect(a.digest).not.toBe(b.digest);
   });
 
   it('handles non-string payloads via JSON.stringify', () => {
@@ -37,7 +37,7 @@ describe('summarizeForLog', () => {
       { role: 'assistant' as const, content: 'hi' },
     ];
     const result = summarizeForLog(messages);
-    expect(result.sha256).toMatch(/^[0-9a-f]{12}$/);
+    expect(result.digest).toMatch(/^[0-9a-f]{16}$/);
     expect(result.len).toBe(JSON.stringify(messages).length);
     // No original content leaks
     expect(JSON.stringify(result)).not.toContain('hello');
