+++ services/platform/convex/lib/response_cache/cache_key.ts
@@ -1,31 +1,8 @@
 import type { GenerationParams } from '../agent_chat/types';
+import { fnv1aHash } from '../fnv1a';
 import { normalizeForCache } from './normalize';
 
 const CACHE_VERSION = 'v1';
-
-/**
- * Simple FNV-1a hash for cache key generation.
- * Not cryptographic — only needs collision resistance for cache deduplication.
- */
-function fnv1aHash(str: string): string {
-  let hash = 0x811c9dc5; // FNV offset basis
-  for (let i = 0; i < str.length; i++) {
-    hash ^= str.charCodeAt(i);
-    hash = Math.imul(hash, 0x01000193); // FNV prime
-  }
-  // Convert to unsigned 32-bit hex + append length-based suffix for extra uniqueness
-  const h1 = (hash >>> 0).toString(16).padStart(8, '0');
-
-  // Second pass with different seed for 64-bit equivalent
-  let hash2 = 0x6c62272e;
-  for (let i = str.length - 1; i >= 0; i--) {
-    hash2 ^= str.charCodeAt(i);
-    hash2 = Math.imul(hash2, 0x01000193);
-  }
-  const h2 = (hash2 >>> 0).toString(16).padStart(8, '0');
-
-  return `${h1}${h2}`;
-}
 
 /**
  * Compute a cache key from stable inputs.
